hoanthien

+++ lab/src/main/java/it/lab/service/ThanhToanService.java
@@ -1,6 +1,8 @@
 package it.lab.service;
 
+import it.lab.common.Email;
 import it.lab.common.ResponObject;
+import it.lab.common.Template;
 import it.lab.dto.DiaChiDTO;
 import it.lab.dto.GioHangDTO;
 import it.lab.dto.PhuongThucThanhToanDTO;
@@ -8,6 +10,7 @@ import it.lab.dto.PhuongThucVanChuyenDTO;
 import it.lab.entity.*;
 import it.lab.enums.APIStatus;
 import it.lab.enums.TrangThaiHoaDon;
+import it.lab.enums.TrangThaiNguoiDungVoucher;
 import it.lab.iservice.IThanhToan;
 import it.lab.modelcustom.request.TaoHoaDonOnline;
 import it.lab.modelcustom.respon.CheckOut;
@@ -21,11 +24,16 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 public class ThanhToanService implements IThanhToan {
     @Autowired
     private GioHangRepo _gioHangRepo;
+    @Autowired
+    private VoucherRepo _voucherRepo;
+    @Autowired
+    private NguoiDungVoucherRepo _nguoiDungVoucherRepo;
     @Autowired
     private NguoiDungRepo _nguoiDungRepo;
     @Autowired
@@ -40,6 +48,23 @@ public class ThanhToanService implements IThanhToan {
     private HoaDonRepo _hoaDonRepo;
     @Autowired
     private HoaDonChiTietRepo _hoaDonChiTietRepo;
+    @Autowired
+    private SanPhamRepo _sanPhamRepo;
+
+    private void thayDoiSoLuongKhiConfirmHoaDon(long hoaDonId) {
+        HoaDon hoaDon = _hoaDonRepo.findById(hoaDonId).get();
+        var chiTiet = _hoaDonChiTietRepo.findHoaDonChiTietsByHoaDon(hoaDon);
+        for (var item : chiTiet) {
+            SanPhamChiTiet sp = item.getSanPhamChiTiet();
+            SanPham sanPham = sp.getSanPham();
+            sanPham.setSoLuongTon(sanPham.getSoLuongTon() - item.getSoLuong());
+            sanPham.setSoLuongDaBan(sanPham.getSoLuongDaBan() + item.getSoLuong());
+            sp.setSoLuongDaBan(sp.getSoLuongDaBan() + item.getSoLuong());
+            sp.setSoLuongTon(sp.getSoLuongTon() - item.getSoLuong());
+            _sanPhamChiTietRepo.save(sp);
+            _sanPhamRepo.save(sanPham);
+        }
+    }
 
     @Override
     public ResponObject<CheckOut, APIStatus> layDuLieuThanhToan(Long nguoiDungId) {
@@ -65,13 +90,22 @@ public class ThanhToanService implements IThanhToan {
         NguoiDung nguoiMua = dc.get().getNguoiDung();
         HoaDon hd = new HoaDon();
         hd.setDiaChiGiao(dc.get());
+        if (yeuCau.getVoucherId() != null) {
+            Voucher v = _voucherRepo.findById(yeuCau.getVoucherId()).get();
+            List<NguoiDungVoucher> lst = _nguoiDungVoucherRepo.findNguoiDungVouchersByNguoiDungAndVoucher(nguoiMua, v)
+                    .stream().filter(x->x.getTrangThai()==TrangThaiNguoiDungVoucher.SUDUNG).collect(Collectors.toList());
+            NguoiDungVoucher ndv = lst.get(0);
+            ndv.setTrangThai(TrangThaiNguoiDungVoucher.DASUDUNG);
+            hd.setVoucherGiam(ndv);
+            _nguoiDungVoucherRepo.save(ndv);
+        }
         hd.setGhiChu(yeuCau.getGhiChu());
         hd.setNgayTao(LocalDateTime.now());
         hd.setNguoiMua(nguoiMua);
         hd.setPhuongThucThanhToan(pttt.get());
         hd.setPhuongThucVanChuyen(ptvc.get());
         hd.setPhiGiaoHang(yeuCau.getPhiVanChuyen());
-        hd.setTrangThai(TrangThaiHoaDon.CHOGIAOHANG);
+        hd.setTrangThai(TrangThaiHoaDon.CHOXACNHAN);
         List<GioHang> ghList = _gioHangRepo.findGioHangsByNguoiMua(nguoiMua);
         Double giaTri = 0d;
         for (GioHang gh : ghList) {
@@ -79,10 +113,31 @@ public class ThanhToanService implements IThanhToan {
             giaTri += sp.getGiaBan() * gh.getSoLuong();
         }
         hd.setGiaTriHd(giaTri + yeuCau.getPhiVanChuyen());
+        //  String thongBao = Template.hoaDonMoi(hd);
+        //    guiThongBaoChoNhanVien(thongBao);
         _hoaDonRepo.save(hd);
-        hd.setHoaDonChiTietList(taoHoaDonChiTiet(ghList,hd.getId()));
+        hd.setMaHoaDon("HD" + hd.getId());
+        _hoaDonRepo.save(hd);
+        hd.setHoaDonChiTietList(taoHoaDonChiTiet(ghList, hd.getId()));
+        thayDoiSoLuongKhiConfirmHoaDon(hd.getId());
         _gioHangRepo.deleteAll(ghList);
         return null;
+    }
+    private void guiThongBaoChoNhanVien(String thongBao) {
+        List<NguoiDung> lst = _nguoiDungRepo.findAll().stream().filter(x -> {
+            for (var item : x.getQuyenNguoiDungList()) {
+                if (item.getQuyen().getId() == 3 || item.getQuyen().getId() == 2) {
+                    return true;
+                }
+            }
+            return false;
+        }).toList();
+        for (var item : lst) {
+            if (item.getEmail() != null) {
+                Email email = new Email();
+                email.sendContentHTML(item.getEmail(), "Hóa đơn mới", thongBao);
+            }
+        }
     }
 
     @Override
@@ -95,6 +150,12 @@ public class ThanhToanService implements IThanhToan {
         }
         NguoiDung nguoiMua = dc.get().getNguoiDung();
         HoaDon hd = new HoaDon();
+        if (yeuCau.getVoucherId() != null) {
+            Voucher v = _voucherRepo.findById(yeuCau.getVoucherId()).get();
+            List<NguoiDungVoucher> lst = _nguoiDungVoucherRepo.findNguoiDungVouchersByNguoiDungAndVoucher(nguoiMua, v)
+                    .stream().filter(x->x.getTrangThai()==TrangThaiNguoiDungVoucher.SUDUNG).collect(Collectors.toList());
+            hd.setVoucherGiam(lst.get(0));
+        }
         hd.setDiaChiGiao(dc.get());
         hd.setGhiChu(yeuCau.getGhiChu());
         hd.setNgayTao(LocalDateTime.now());
@@ -102,7 +163,7 @@ public class ThanhToanService implements IThanhToan {
         hd.setPhuongThucThanhToan(pttt.get());
         hd.setPhuongThucVanChuyen(ptvc.get());
         hd.setPhiGiaoHang(yeuCau.getPhiVanChuyen());
-        hd.setTrangThai(TrangThaiHoaDon.CHOGIAOHANG);
+        hd.setTrangThai(TrangThaiHoaDon.CHOTHANHTOANBANKING);
         List<GioHang> ghList = _gioHangRepo.findGioHangsByNguoiMua(nguoiMua);
         Double giaTri = 0d;
         for (GioHang gh : ghList) {
@@ -110,13 +171,36 @@ public class ThanhToanService implements IThanhToan {
             giaTri += sp.getGiaBan() * gh.getSoLuong();
         }
         hd.setGiaTriHd(giaTri + yeuCau.getPhiVanChuyen());
-
         _hoaDonRepo.save(hd);
         hd.setMaHoaDon("HD" + hd.getId());
-        hd.setHoaDonChiTietList(taoHoaDonChiTiet(ghList,hd.getId()));
+        hd.setHoaDonChiTietList(taoHoaDonChiTiet(ghList, hd.getId()));
         _hoaDonRepo.save(hd);
-        _gioHangRepo.deleteAll(ghList);
+        //  _gioHangRepo.deleteAll(ghList);
+        //  thayDoiSoLuongKhiConfirmHoaDon(hd.getId());
         return hd.getMaHoaDon();
+    }
+
+    @Override
+    public void checkThanhToanVNPay(String maHD, Long trangThai) {
+        var hoaDon = _hoaDonRepo.findHoaDonByMaHoaDon(maHD);
+        if (!hoaDon.isPresent()) {
+            return;
+        }
+        if (trangThai == 0) {
+            List<GioHang> ghList = _gioHangRepo.findGioHangsByNguoiMua(hoaDon.get().getNguoiMua());
+            _gioHangRepo.deleteAll(ghList);
+            thayDoiSoLuongKhiConfirmHoaDon(hoaDon.get().getId());
+            hoaDon.get().setNgayThanhToan(LocalDateTime.now());
+            if (hoaDon.get().getVoucherGiam() != null) {
+                hoaDon.get().getVoucherGiam().setTrangThai(TrangThaiNguoiDungVoucher.DASUDUNG);
+                _nguoiDungVoucherRepo.save(hoaDon.get().getVoucherGiam());
+            }
+            hoaDon.get().setTrangThai(TrangThaiHoaDon.CHOGIAOHANG);
+            _hoaDonRepo.save(hoaDon.get());
+        } else {
+            hoaDon.get().setTrangThai(TrangThaiHoaDon.CUAHANGHUY);
+            _hoaDonRepo.save(hoaDon.get());
+        }
     }
 
     private List<HoaDonChiTiet> taoHoaDonChiTiet(List<GioHang> gioHangList, Long hoaDonId) {
@@ -124,9 +208,8 @@ public class ThanhToanService implements IThanhToan {
         HoaDon hoaDon = _hoaDonRepo.findById(hoaDonId).get();
         for (GioHang gh : gioHangList) {
             SanPhamChiTiet spct = _sanPhamChiTietRepo.findById(gh.getSanPhamChiTiet().getId()).get();
-            spct.setSoLuongTon(spct.getSoLuongTon() - gh.getSoLuong());
             HoaDonChiTiet hoaDonChiTiet = new HoaDonChiTiet();
-            hoaDonChiTiet.setNgayTao(LocalDate.now());
+            hoaDonChiTiet.setNgayTao(LocalDateTime.now());
             hoaDonChiTiet.setSanPhamChiTiet(spct);
             hoaDonChiTiet.setSoLuong(gh.getSoLuong());
             hoaDonChiTiet.setDonGia(gh.getSanPhamChiTiet().getGiaBan());
